fix(ui5-input): fire change when value is mutated via API
Firing change event used to rely on the user input and did not
consider changes through the "value" property. Now such changes are taken into count.
In addition, the change event is also fired whenever suggestion item is selected.
The issue can be observed in the ui5-datepicker. When the user types "tomorrow" and presses ENTER,
everything is fine, the value in the input field changes to the particular date.
But, when the user types "tomorrow" again - nothing happens, as the internal ui5-input does not fire "change" event.
Fixes: https://github.com/SAP/ui5-webcomponents/issues/154