refactor: make custom CSS theme independent (#386)
With the move to CSS custom properties, the CSS that is added to the shadow DOM is theme independent and never changed when the theme is changed. Theme dependent styling has to be achieved by use of CSS cutsom properties only. This method is adjusted to reflect the standard framework behaviour.

BREAKING CHANGE: the signature of the addCustomCSS method exported by "@ui5/webcomponents-base/Theming.js" is changed from addCustomCSS(tag, theme, css) to addCustomCSS(tag, css)