fix: fix API build for child components
Some of the components do not have separate sample page, but are appended to others sample page.
The API of those components is part (in most cases) of parent components API, that use them as composite components.
Issue: the API of the composite/child components misses base properties.
Root cause: the API generataion of the parent component is done before the API generation of the composite/child compoenents.
Solution: calculate the API of those composite components before their addiotion to parent`s component API.
FIXES: https://github.com/SAP/ui5-webcomponents/issues/390