feat(ui5-toast): WebComponent is enriched with ACC and tests
Toast now have accessibility enabled.
The text inside it is red right after Toast is opened.
This is achieved by adding a role="alert" on the "ui5-toast-root".
Also in order to make it work, we are now using display none/flex on
the parent instead of visibility hidden/visible as it was previously.

There is a slight refactoring on the logic of the open property toggling.
We now always set it to true after an animation frame.

Tests are added in the Toast.spec.js file.

Toast.html is slightly altered in order to make testing easier.

Toast.sample.html is slightly altered in order to demonstrate the default
placement property ("BottomCenter"). This is achieved by removing the
explicit placement attribute on one of the Toasts.