feat(ui5-toast): improved samples, styling, animation behaviour, tests
Sample changes:
More samples are added in the playground, which demonstrate default
Toast, as well as some short/long duration options for the component.
Also the script for the handlers in now simplified/shortened in both
playground and test pages.

Style changes:
Added "--_ui5_toast_background" and "--_ui5_toast_shadow" CSS parameters.
Font-family, color, and font-size of the "ui5-toast-text" are now on the
host.
These changes give the web-component user, the freedom to alter
the styling of the Toast.

CSS class name changes:
The "ui5-toast-root-parent" class is now renamed to "ui5-toast-root".
The "ui5-toast-root" class is now called "ui5-toast-content".

Animation changes:
The transition (fade out animation) is now handled differently.
It is still a third of the Toast's duration property, but if
it happens to be more than 1000ms, we force it to be 1000ms.
This way we omit prolonged fade out times, when
the duration property is longer, than 3000ms (default).

Tests changes:
Spec files are now aligned with the new transition behaviour.
Also some new tests are now added in order to cover mose cases.