fix(ui5-select): fix component baseline alignment (#1075)
FIXES: https://github.com/SAP/ui5-webcomponents/issues/1074

**Background**
The `overflow: hidden` causes the ui5-select to be pushed upwards the main baseline (as described in the issue). Originally, the style was introduced  not to allow the hover of the arrow to cover the input border. Now, this is resolved by moving this `overflow: hidden` to inner element and in the same time not disrupting the baseline alignment of the whole component.