WIP(ui5-upload-collection): implement new webcomponent
ui5-upload-collection is new webcomponent located in fiori package,
which benefits and brings some of ui5-list features, but it shows
different items of type ui5-upload-collection-item.
It is in the responsibility of the app developer to put ui5-file-uploader
in the header, create and add new ui5-upload-collection-item(s) to the collection.

Main features:
- header slot
- items slot (default)
- items have special visualisation, suitable for files representation
- items have "file" property, to allow app developers to set and get file object
- drag and drop overlay