feat(ui5-datetime-picker): introduce new component
Background
The DateTimePicker component alows users to select both date (day, month and year) and time (hours, minutes and seconds).
The component opens a popover with calendar and time controls.
On smaller screen and on phone device there is a switch to change between the calendar and time controls.
Technical points
- extends DatePicker
- reuses the DatePicker template and WheelSlider component
- refactors the DatePicker to make it more extensible and overwrite few methods
- extracts time related functions for common usage by the DateTimePicker and the TimePicker