refactor(ui5-input): add template hook (#1608)
Adds a template hook to allow components, extending the Input to add content, positioned right before the input field. The use case is an ui5-input, used as a search field that needs to show a "transitive search" info, the category that the users search in. For more info see the related issue.

Related to: #1588