chore: fix btn slot display
Recent fix for how icon-only button with just comment nodes is rendered, has affectedthe the rendering of images, avatars inside the button.
This can be seen in the ShellBar page, where the ShellBar uses ui5-avatar for its profile, internally rendered in a ui5-button. Currently the ui5-avatar is not visible as display: none is applied due to the icon-only state of the ui5-button. Now we filter the comment nodes and check for the childNodes length to determine if its icon-only button or not.