chore: fix btn slot display (#1708)
Recent fix for how icon-only button with just comment nodes is rendered, has affected rendering of images, avatars inside the button.
This can be seen in the ShellBar page, where ui5-avatar is used for profile slot, internally slotted in a ui5-button. The profile image is not visible as "display: none" is applied due to the icon-only state of the ui5-button.
Now we filter the comment nodes and check for the childNodes length to determine if its icon-only button or not.