chore(ui5-input): set data-key to group item
There is data-key set on the standard suggestions, but it's missing on the group suggestion. It is used for the mouseover/out events. Curerntly, if the users hovers on a group item, error is thrown as the corresponding suggestion item cant be found without this data-key attribute. Other way would be to not fire the events for hovoring over group items, but at the end I think this is better.