refactor(ui5-input): Add hook for value update on preview (#1956)
Added extension point to allow extending components control the behaviour on suggestion preview. The default behaviour of the Input with suggestions is to replace the current value with the value of the previewed suggestion, when the user navigates with UP/DOWN arrows. But, there are use cases as in the attached issue, that instead of replacing the entire text, just part of it has to be replaced, changed or added to.

FIXES: #1878