fix(components): fix setting contentDensity styles
Fixes
- TimePicker was never considering contentDensity
- DatePicker and DateTimePicker used to call updateStaticAreaItemContentDensity before the staticAreaItemDom is available

Solution:
handle setting of contentDensity on framework level when possible and remove updateStaticAreaItemContentDensity calls from ShellBar, ComboBox, Select, Input, MultiCombobox. The updateStaticAreaItemContentDensity method exists and used in the TabContainer only, because the component does not invalidate on open/close of the overflow and needs to force the contentDensity update.