refactor(framework):  fetch default language (#2175)
After the change all languages would be fetched over the network, including the defaultLanguage, that we have inlined. This will be the default behaviour. However, there is a new configuration option to stop fetching the default language, called "useDefaultLanguage". When the "useDefaultLanguage" is enabled,
the inlined texts will be used instead.

Related to: https://github.com/SAP/ui5-webcomponents/issues/2169