chore(ui5-fcl): add template hooks for toggle arrows
Add template hooks for the toggle arrows of thew FlexibleColumnLayout,
so that a derived class can fully customize them.
For example: provide different or new ARIA attribbutes or change the values of existing ones.
Note: We considered providing slots or proeprties,
but we wanted to keep the API more simple and aggreed on this solution.

FIXES: https://github.com/SAP/ui5-webcomponents/issues/2183