Add state storage for tracking state changes
In order to always keep the visual UI representation and the internal
state in sync, the component has a 'state storage' that is updated when
the current state is changed due to a user action.

It's used do differentiate when a property has not been changed via user
interaction but programatically - because the previous state is always
updated in the interaction handlers.

The initial visual UI representation synchronization with the internal
state is also managed via the previously saved 'state storage', as in
that case there is no values yet stored.