fix(ui5-li): fix title update when initially empty (#2362)
When the textContent is an empty string initially, the following element is not rendered
<span part="title" class="ui5-li-title"><slot></slot></span>
Later when we update the textContent, it is not displayed as it relies on the slotting, but the slot element is not there. The issue has been found in a specific Select use case, referenced below. I considered making a ui5-select-list-item and make that change in a separate template, but I can't find side effects or visual degradations.

FIXES https://github.com/SAP/ui5-webcomponents/issues/2342