Add additional validations and normalizations, refactoring
Normalization of the min and max value is introduced. If the min
property is set to be more than the max one they are swapped.

Validation of the step property is improved. If a negative value is
provided it gets converted to its positive equivalent.
Warning are now being display if the step value is 0, a negative number
or not a valid floating number.

tickmarks property is renamed to showTickmarks

Some refactoring is performed according to the code reviews.