feat(ui5-avatar-group): initial implementation
Introduce new component in the "main" package, called AvatarGroup. The component is used to
display a group of avatars arranged horizontally. It is useful to visually
showcase a group of related avatars, such as, project team members or employees.
The control allows you to display the avatars in different sizes,
depending on your use case.
Supports 2 types - "Group" and "Invidual":
Group type: The avatars are displayed as partially overlapped on
top of each other and the entire group has one click/tap area.

Individual type: The avatars are displayed side-by-side and each
avatar has its own click/tap area.

Fixes: https://github.com/SAP/ui5-webcomponents/issues/2409