Add more tests, add cozy styles, fix input event
New tests are added to cover using 'effective' properties for
calculation instead of normaliziations of the real ones.

Added cozy styles and set as default while the compact css vars are now
moved to sizes-parameters.css files.

Input event is now fixed and fired only after the current value is
changed by user interaction.

Updating the labels based on the other properties is fixed - labels are
now getting re-created if the one of the other component's properties is updated.

Tickmarks bug when sometimes the tickmars are hidden under the parent's
background is fixed, the tickmark DOM element no longer has a z-index
set to -1.