feat(ui5-multi-combobox): support two column layout for items (#2642)
Enables two-column layout items in the MultiComboBox. The MultiComboBoxItem already inherits the additionalText property from the ComboBoxItem, but never used it. The property is visible in the API reference and misleading users that can have two-column layout in the MultiComboBox as well. The change just uses the property in the template and forwards it to the physically rendered list item.

FIXES: #2637