fix(ui5-input): fix Backspace on Input type=Number
The "_handleInput" handler works with event from type "InputEvent" that never has "key" or "keyCode", that's why "isBackspace" or any other checks regarding pressed keys will always result in false. The solution is to execute the logic on keydown where the "KeyboardEvent" has "key" or "keyCode", store the result and later use it in "_handleInput".

FIXES: https://github.com/SAP/ui5-webcomponents/issues/2841