fix(ui5-page): correct background-design styles (#2862)
The styles used to rely on the property name backgroundDesign, which is not correct as in the DOM we use the attribute name background-design. In addition, the "Standard" background design is removed, because it's the same as "Solid".
As a result the background Design default value is changed from "Standard" to "Solid".

BREAKING_CHANGE: The backgroundDesign default value is changed from "Standard" to "Solid".