fix(ui5-wizard): prevent scrolling to step's very top (#2888)
On re-rendering the Wizard auto scrolls to the currently selected item's top, which is useful when the step selection is changed programatically. But, when re-rendering occurs for some other reason and the step selection remains the same, the Wizard still scrolls to the top of the selected step, which is not expected. Now, this is fixed by checking if the selection has been changed.

Fixes: #2883