refactor(ui5-wizard): configure step change position
Background
There si complains that if the user is at the very beginning of a step and scrolls a bit upwards, the selection switches to the previous step, although most of the viewport is taken by the first one.

Solution
Add private/protected API, called step-switch-threshold
to give more control of the point the switch happens.
It support float values between 0.5 and 1, where 0.5 means the step will be switched after the user scrolls half of the current step, and 1 - when the current step is entirely scrolled out. The default value is 0.7, which is very close to the behaviour of the sap.m.Wizard.