feat(ui5-avatar-group): new slot overflowButton (#3037)
Added new slot which allows to add custom overflow button and "overflow" event to let you know when avatars
hide/show in order to update the overflow button text. If not provided, the AvatarGroup will display the built-in overflow button. The slot is provided to allow you defining the text of the overflow button and let you set additional aria attributes, such as aria-label.

Fixes: #2912