fix(ui5-input): inner input padding is correctly updated (#3015)
Prevously, after padding: 0 was applied when the condition is false, it remained 0. That could be easily reproduced with native HTML as well. Setting 'undefined' as a value does not change the padding to the one specified in the component's css styles.
https://snippix.only.sap/snippets/52143
Now, when the padding should not be 0, the inline style is removed, instead of setting 'undefined' as a value

FIXES: #2940