fix: keep invisible texts out of the view port (#3085)
Add negative coordinates to the a11y invisible texts to ensure they remain out of the view port. Already this is the case in most of the browsers, but on IE they can be still observed. The negative coordinates do not affect the a11y and do not prevent Screen Readers from announcing the texts.

FIXES: #2723
CLOSES: #2723