fix(ui5-list): prevent load-more on initial intersection (#3105)
This is a change already merged in release-0.29, because of a tight deadline, set by stakeholders. Now we ensure that the issue is fixed in the future versions. The commit in the release-0.29 branch is the following: af2070f

And the change fixes the case when the List does not actually overflow, the bottom most part of the List is visible and the IntersectionObserver detects that the bottom of the List is reached. What happens in the apps is that the List fires "load-more" just before the items that the app wants to show initially are actually rendered, which results in additional loading, not due to user scroll.