chore(ui5-table): forward focus correctly to row element
If the user clicks on a non-focusable element within the ui5-table-cell, the focus goes to the body or to the first focusable element above, so we have to bring it back to the row. If the user clicks on input, button or a similar clickable element, the focus remains on that element.