refactor(ui5-wizard): selection-change renamed to step-change
We have decided to rename the selection-change event in order for its name to better represent its origins.
In relation to #3107.

BREAKING_CHANGE: The “selection-change” event has been renamed to "step-change". In addition, the event param "selectedStep" has been renamed to "step" and the event param "previouslySelectedStep" has been renamed to "previousStep".