feat(ui5-input): introduce new SuggestionGroupItem (#3248)
Introduce a dedicated component SuggestionGroupItem to split the suggestions into groups. Previously the app developers had to use the standard SuggestionItem and set its "group" property. But, as the group item do not have most of the interactive features, we decided to introduce dedicated component for the purpose. Related to: #3107

BREAKING_CHANGE: The "group" property of the SuggestionItem is removed, use the SuggestionGroupItem component to create grouping as follows:

```html
<ui5-input show-suggestions>
  <ui5-suggestion-group-item text="Group #1"></ui5-suggestion-group-item>
  <ui5-suggestion-item text="Item #1"></ui5-suggestion-item>
  <ui5-suggestion-item text="Item #2"></ui5-suggestion-item>
</ui5-input>
```