refactor(ui5-color-palette): change API (#3420)
As part of the #3107 we decided to introduce a new component ColorPalettePopover and to remove most of the ColorPalette API (more/recentColors features). With this change the ColorPalette basically supports only the colors slot (as the sap.m.ColorPalette control does).

BREAKING_CHANGE: The showMoreColors property is removed, ColorPalettePopover component will be introduced in future and provide this functionality
BREAKING_CHANGE: The showRecentColors property is removed, ColorPalettePopover component will be introduced in future and provide this functionality
BREAKING_CHANGE: The value property has been removed.