refactor(ui5-color-palette-item): stableDomRef is obsolete (#3438)
The stableDomRef is meant for components that don't have a template on their own, but are pure elements, that the parent component renders (for example Select and Options). In this case the ColorPaletteItem is a component with own template and does not need to expose a special property like this to app developers to get its DOM ref as there is already getDomRef method that is sufficient.

BREAKING_CHANGE: stableDomRef is removed, if you need to get the item's DOM ref, use getDomRef on the item instance.