feat(ui5-combobox): add suggestions grouping (#3469)
Enable users to set items to groups. Groups are visually identified in the suggestions list by group headers.
For group headers a new component is introduced - ComboBoxGroupItem. The grouping is based on the items order as they are declared in the markup - the items between two group headers are considered to belong to the first one.

Fixes: #3371