fix(ui5-combobox): Double announcement on key navigation fixed
While navigating with the keyboard over the list of items in the ComboBox, some of the screen readers anounce the currently seleted value twice.
The reasoning behind this behaviour is that the value of ComboBox's input gets updated and then it gets highlighted. The issue is that the value of the input is updated asynchronously and then the highlyghting is again executed in an async manner. In essence, the situation is:
``
setTimeout(() => input.value = newValue, 0);
setTimeout(() => input.setSelectionRange(0, newValue.length), 0);
``

Value update is async, but discreet as the handlebars/webcomponents' renderer needs to update the value as it is bound to `this.value` property. The highlighting is quite clear- the code is put in setTimeout. This timeout is needed as the renderer has not yet updated the DOM and this ensures that the highlight would be executed after input's value is the recent one.