fix(ui5-responsive-popover): fixed position of downward pointing arrow
ResponsivePopover resets the custom property `_ui5_popup_content_padding`
as it is used by other components.

The calculation for the position of the arrow depended on the padding,
and the calculation will be wrong when the padding is set to "0".

Solved by changing the positioning of the arrow by using `top`.

Fixes #3500