fix(ui5-checkbox): add aria-hidden attribute to icon (#3511)
The role presentation on the icon SVG is ignored in some cases:
*The element is focusable, e.g. it is natively focusable like an HTML link or input, or it has a tabindex attribute.
*The element has any of the twenty-one global ARIA states and properties, e.g., aria-label.

With this change we introduce ariaHidden property on ui5-icon to be able to completly hide it from accessibility tree mapping.
Also, change the default value of effectiveAccessibleName to undefined as a value of "" will still set the aria-label attribute and the presentation role will be ignroed.

Fixes #3433