fix(ui5-combobox): Fix value reset on ESC #3886
When the picker is closed the entered value is not cleared. When the picker is opened, it closes but the value remains in the input field. Now the component is synced with the desired by specifications behavior.
If the suggestions popover is opened: ESC closes it.
If the suggestions are not opened: ESC resets the value to the last confirmed by the user with 'ENTER' or item click, or if there isn't one - with the initial value of the component when focused by the user.

Fixes #3784
Closes #3784