fix(ui5-list): change role mappings so no interactive elements are ne… (#3952)
Fixes: #3927

Issue: Interactive elements should not be nested inside element with role option.
There are nested interactive elements inside list items in selection mode - ui5-radio-button & ui5-checkbox.

Summary of changes:

the role mapping is changed from listbox -> option[] to list -> listitem[];
the selected state announcement is handled separately with aria-describedby as the listitem role does not support aria-selected;
aria-multiselectable is removed because it is not supported on the role list;
Note: aria-labelledby should not be used for the "Selected" & "Not Selected" texts as NVDA reads the whole value of the attribute when some of its references changes content. In this case - the whole content of the item.